pyproject.toml
[tool.poetry]
name = "hustlers-den"
version = "0.1.0"
description = "Knowledge Resource Aggregator for Teams"
authors = ["Abhishek Juneja <abhishek.juneja145@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
appnope = "0.1.0"
astroid = "1.6.5"
backcall = "0.1.0"
certifi = "2018.4.16"
chardet = "3.0.4"
coreapi = "2.3.3"
coreschema = "0.0.4"
decorator = "4.3.0"
Django = "2.2.12"
django-debug-toolbar = "1.9.1"
django-extensions = "^2.2.9"
django-filter = "^2.2.0"
django-jet = "1.0.7"
django-querycount = "0.7.0"
djangorestframework = "^3.11.0"
djangorestframework-jwt = "1.11.0"
graphviz = "0.8.4"
idna = "2.7"
ipaddress = "1.0.22"
ipython = "6.4.0"
ipython-genutils = "0.2.0"
isort = "4.3.4"
itypes = "1.1.0"
jedi = "0.12.0"
Jinja2 = "^2.11.2"
lazy-object-proxy = "1.3.1"
MarkupSafe = "1.0"
mccabe = "0.6.1"
openapi-codec = "1.3.2"
parso = "0.2.1"
pexpect = "4.6.0"
pickleshare = "0.7.4"
prompt-toolkit = "1.0.15"
ptyprocess = "0.5.2"
Pygments = "2.2.0"
PyJWT = "1.6.4"
pylint = "1.9.2"
python-dateutil = "2.7.3"
pytz = "2018.4"
raven = "6.9.0"
requests = "^2.23.0"
simplegeneric = "0.8.1"
simplejson = "3.15.0"
six = "1.11.0"
sqlparse = "0.2.4"
traitlets = "4.3.2"
uritemplate = "3.0.0"
urllib3 = "^1.25.9"
wcwidth = "0.1.7"
wrapt = "1.10.11"
celery = "^4.4.2"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.5"
whitenoise = "^5.1.0"
Werkzeug = "^1.0.1"
gunicorn = "^20.0.4"
drf-yasg = {extras = ["validation"], version = "^1.17.1"}
pre-commit = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
